+++ test/json.js
@@ -285,26 +285,47 @@ describe('bodyParser.json()', function(){
     })
   })
 
-  it('should support utf-8', function(done){
-    var server = createServer()
+  describe('charset', function(){
+    var server;
+    before(function(){
+      server = createServer()
+    })
 
-    request(server)
-    .post('/')
-    .set('Content-Type', 'application/json; charset=utf-8')
-    .send('{"name":"论"}')
-    .expect(200, '{"name":"论"}', done);
-  })
+    it('should parse utf-8', function(done){
+      var test = request(server).post('/')
+      test.set('Content-Type', 'application/json; charset=utf-8')
+      test.write(new Buffer('7b226e616d65223a22e8aeba227d', 'hex'))
+      test.expect(200, '{"name":"论"}', done)
+    })
 
-  it('should support {"test":"å"}', function (done) {
-    // https://github.com/visionmedia/express/issues/1816
-    var server = createServer();
+    it('should parse utf-16', function(done){
+      var test = request(server).post('/')
+      test.set('Content-Type', 'application/json; charset=utf-16')
+      test.write(new Buffer('feff007b0022006e0061006d00650022003a00228bba0022007d', 'hex'))
+      test.expect(200, '{"name":"论"}', done)
+    })
 
-    request(server)
-    .post('/')
-    .set('Content-Type', 'application/json; charset=utf-8')
-    .set('Content-Length', '13')
-    .send('{"test":"å"}')
-    .expect(200, '{"test":"å"}', done);
+    it('should parse when content-length != char length', function(done){
+      var test = request(server).post('/')
+      test.set('Content-Type', 'application/json; charset=utf-8')
+      test.set('Content-Length', '13')
+      test.write(new Buffer('7b2274657374223a22c3a5227d', 'hex'))
+      test.expect(200, '{"test":"å"}', done)
+    })
+
+    it('should default to utf-8', function(done){
+      var test = request(server).post('/')
+      test.set('Content-Type', 'application/json')
+      test.write(new Buffer('7b226e616d65223a22e8aeba227d', 'hex'))
+      test.expect(200, '{"name":"论"}', done)
+    })
+
+    it('should fail on unknown charset', function(done){
+      var test = request(server).post('/')
+      test.set('Content-Type', 'application/json; charset=koi8-r')
+      test.write(new Buffer('7b226e616d65223a22cec5d4227d', 'hex'))
+      test.expect(415, 'unsupported charset', done)
+    })
   })
 })